Use Deface frontend overrides only for test environment
Deface overrides for frontend pages are by default injected only in test
environment in order to make specs pass and as a basic working example of how
the plugin may work in real life.

Some usage instructions have been added to README in order to introduce the
change and show the plugin views usage in a real Solidus app.